+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     'pallets/celer-pay',
     'pallets/celer-pay/rpc',
     'pallets/celer-pay/rpc/runtime-api',
+    'pallets/condition-caller',
     'pallets/celer-contracts',
     'pallets/celer-contracts/rpc',
     'pallets/celer-contracts/rpc/runtime-api',
