+++ pallets/celer-pay/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "celer-pay-module"
-version = "0.8.8"
+version = "0.8.9"
 authors = ["Celer"]
 edition = "2018"
 license = "MIT"
@@ -22,13 +22,14 @@ frame-support = { version = "2.0.0", default_features = false }
 frame-system = { version = "2.0.0", default_features = false }
 pallet-balances = { version = "2.0.0", default_features = false }
 pallet-timestamp = { version = "2.0.0", default_features = false }
-mock-numeric-condition = { path = "../mock-numeric-condition", version = "0.8.8", default_features = false }
-mock-boolean-condition = { path = "../mock-boolean-condition", version = "0.8.8", default_features = false }
-celer-pay-module-rpc-runtime-api = { path = "./rpc/runtime-api", version = "0.8.8", default_features = false }
+runtime-module-condition-caller = { path = "../condition-caller", version = "0.8.9", default_features = false }
+celer-pay-module-rpc-runtime-api = { path = "./rpc/runtime-api", version = "0.8.9", default_features = false }
 celer-contracts = { path = "../celer-contracts", version = "0.8.8", default_features = false }
 
 [dev-dependencies]
 pallet-randomness-collective-flip = { version = "2.0.0", default_features = false }
+mock-numeric-condition = { path = "../mock-numeric-condition", version = "0.8.9", default_features = false }
+mock-boolean-condition = { path = "../mock-boolean-condition", version = "0.8.9", default_features = false }
 
 [features]
 default = ["std"]
@@ -43,8 +44,7 @@ std = [
 	"frame-system/std",
 	"pallet-balances/std",
 	"pallet-timestamp/std",
-	"mock-numeric-condition/std",
-	"mock-boolean-condition/std",
+	"runtime-module-condition-caller/std",
 	"celer-pay-module-rpc-runtime-api/std",
 	"celer-contracts/std",
 ]
