+++ pallets/celer-pay/rpc/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "celer-pay-module-rpc"
-version = "0.8.8"
+version = "0.8.9"
 authors = ['Celer']
 edition = "2018"
 license = "MIT"
@@ -22,7 +22,7 @@ sp-rpc = { version = "2.0.0" }
 serde = { version = "1.0.101", features = ["derive"] }
 sp-runtime = { version = "2.0.0" }
 sp-api = { version = "2.0.0" }
-celer-pay-module-rpc-runtime-api = { version = "0.8.8", path = "./runtime-api" }
+celer-pay-module-rpc-runtime-api = { version = "0.8.9", path = "./runtime-api" }
 
 [dev-dependencies]
 serde_json = "1.0.41"
