+++ pallets/celer-pay/rpc/runtime-api/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "celer-pay-module-rpc-runtime-api"
-version = "0.8.8"
+version = "0.8.9"
 authors = ["Celer"]
 edition = "2018"
 license = "MIT"
