+++ pallets/celer-pay/src/mock.rs
@@ -171,6 +171,8 @@ impl mock_boolean_condition::Trait for TestRuntime {}
 
 impl mock_numeric_condition::Trait for TestRuntime {}
 
+impl runtime_module_condition_caller::Trait for TestRuntime {}
+
 impl Trait for TestRuntime {
     type Currency = pallet_balances::Module<Self>;
     type Event = TestEvent;
