+++ pallets/condition-caller/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "runtime-module-condition-caller"
+version = "0.8.9"
+authors = ["Celer"]
+edition = "2018"
+license = "MIT"
+repository = "https://github.com/celer-network/cChannel-substrate"
+description = "A module for registering runtime module condition and call query API"
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "1.3.4", features = ["derive"], default-features = false }
+sp-std = { version = "2.0.0", default_features = false }
+sp-runtime = { version = "2.0.0", default_features = false }
+frame-support = { version = "2.0.0", default_features = false }
+frame-system = { version = "2.0.0", default_features = false }
+mock-numeric-condition = { path = "../mock-numeric-condition", version = "0.8.9", default_features = false }
+mock-boolean-condition = { path = "../mock-boolean-condition", version = "0.8.9", default_features = false }
+
+[features]
+default = ["std"]
+std = [
+	"codec/std",
+    "sp-std/std",
+	"sp-runtime/std",
+	"frame-support/std",
+	"frame-system/std",
+    "mock-numeric-condition/std",
+	"mock-boolean-condition/std",
+]
