+++ runtime/Cargo.toml
@@ -35,6 +35,7 @@ sp-session = { version = "2.0.0", default-features = false }
 sp-transaction-pool = { version = "2.0.0", default-features = false }
 sp-version = { version = "2.0.0", default-features = false }
 celer-pay-module = { path = "../pallets/celer-pay", default-features = false }
+runtime-module-condition-caller = { path = "../pallets/condition-caller", default-features = false }
 mock-boolean-condition = { path = "../pallets/mock-boolean-condition", default-features = false }
 mock-numeric-condition = { path = "../pallets/mock-numeric-condition", default-features = false }
 celer-contracts = { path = "../pallets/celer-contracts", version = "0.8.8", default-features = false }
@@ -83,6 +84,7 @@ std = [
 	"frame-system/std",
 	"frame-system-rpc-runtime-api/std",
 	"celer-pay-module/std",
+	"runtime-module-condition-caller/std",
     "mock-boolean-condition/std",
     "mock-numeric-condition/std",
 	"celer-contracts/std",
